Student Service/backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.21
psycopg[binary]==3.2.13
python-dotenv==1.0.0
PyJWT==2.8.0
httpx==0.25.0
